Refactor Unit14 Mastering ARM CortexM3-4 folder structure for clarity
- Organized and renamed subfolders into clear sequential topics.
- Renamed simulation results for consistency.
