+++ Unit14 Mastering ARM CortexM3-4/03_Switch_PAL_NPAL_Mode/HAL Drivers/STM32F103x6.h
